--- README.md
@@ -1,6 +1,0 @@
-# Twitter-Lupus-tracker
-I am tracking tweets related to lupus to see how big of a community this disease has and how active the lupus community is on twitter.
-
-Work in Progress:
-Currently working on graphing tweets from the twitter streaming api by frequency per day and location within the USA for a set time frame.
-
