--- python_code/chart.html
@@ -1,20 +1,0 @@
-<html>  
-<head>    
-    <title>Vega Scaffold</title>
-    <script src="http://d3js.org/d3.v3.min.js" charset="utf-8"></script>
-    <script src="http://d3js.org/topojson.v1.min.js"></script>
-    <script src="http://d3js.org/d3.geo.projection.v0.min.js" charset="utf-8"></script>
-    <script src="http://trifacta.github.com/vega/vega.js"></script>
-</head>
-<body>
-    <div id="vis"></div>
-</body>
-<script type="text/javascript">
-// parse a spec and create a visualization view
-function parse(spec)
-{
-vg.parse.spec(spec, function(chart) { chart({el:"#vis"}).update(); });
-}
-parse("time_chart.json");
-</script>
-</html>
